# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/stores": {
      "get": {
        "operationId": "StoreController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "페이지 번호",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지당 항목 수",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "상점 이름 검색",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "카테고리 ID로 필터링",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "ownerId",
            "required": false,
            "in": "query",
            "description": "점주 ID로 필터링",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "활성 상태로 필터링",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isVerified",
            "required": false,
            "in": "query",
            "description": "인증 상태로 필터링",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "정렬 기준",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "name",
                "categoryId"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "정렬 방향",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "상점 목록 반환"
          }
        },
        "summary": "상점 목록 조회",
        "tags": [
          "상점"
        ]
      },
      "post": {
        "operationId": "StoreController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "상점 등록 성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증되지 않은 요청"
          },
          "403": {
            "description": "권한 없음"
          },
          "409": {
            "description": "사업자 등록 번호 중복"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "상점 등록",
        "tags": [
          "상점"
        ]
      }
    },
    "/api/stores/{id}": {
      "get": {
        "operationId": "StoreController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "상점 정보 반환"
          },
          "404": {
            "description": "상점을 찾을 수 없음"
          }
        },
        "summary": "상점 상세 조회",
        "tags": [
          "상점"
        ]
      },
      "patch": {
        "operationId": "StoreController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "상점 수정 성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증되지 않은 요청"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "상점을 찾을 수 없음"
          },
          "409": {
            "description": "사업자 등록 번호 중복"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "상점 정보 수정",
        "tags": [
          "상점"
        ]
      },
      "delete": {
        "operationId": "StoreController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "상점 삭제 성공"
          },
          "401": {
            "description": "인증되지 않은 요청"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "상점을 찾을 수 없음"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "상점 삭제",
        "tags": [
          "상점"
        ]
      }
    },
    "/api/stores/owner/my-stores": {
      "get": {
        "operationId": "StoreController_getMyStores",
        "parameters": [],
        "responses": {
          "200": {
            "description": "내 상점 목록 반환"
          },
          "401": {
            "description": "인증되지 않은 요청"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "내 상점 목록 조회",
        "tags": [
          "상점"
        ]
      }
    },
    "/api/stores/{id}/verify": {
      "patch": {
        "operationId": "StoreController_verifyStore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "인증 상태 변경 성공"
          },
          "401": {
            "description": "인증되지 않은 요청"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "상점을 찾을 수 없음"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "상점 인증 상태 변경 (관리자 전용)",
        "tags": [
          "상점"
        ]
      }
    },
...
    "/api/menu-categories/store/{storeId}": {
      "get": {
        "operationId": "MenuCategoryController_findByStoreId",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "매장별 메뉴 카테고리 목록을 반환합니다."
          }
        },
        "summary": "매장별 메뉴 카테고리 조회",
        "tags": [
          "메뉴 카테고리"
        ]
      },
      "post": {
        "operationId": "MenuCategoryController_create",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMenuCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "메뉴 카테고리가 생성되었습니다."
          },
          "403": {
            "description": "해당 매장에 대한 권한이 없습니다."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "메뉴 카테고리 생성",
        "tags": [
          "메뉴 카테고리"
        ]
      }
    },
...
    "/api/menu-items/store/{storeId}": {
      "get": {
        "operationId": "MenuItemController_findByStoreId",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "매장별 메뉴 아이템 목록을 반환합니다."
          }
        },
        "summary": "매장별 메뉴 아이템 조회",
        "tags": [
          "메뉴 아이템"
        ]
      },
      "post": {
        "operationId": "MenuItemController_create",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateMenuItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "메뉴 아이템이 생성되었습니다."
          },
          "403": {
            "description": "해당 매장에 대한 권한이 없습니다."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "메뉴 아이템 생성",
        "tags": [
          "메뉴 아이템"
        ]
      }
    },
...
    "/api/menu-items/popular/store/{storeId}": {
      "get": {
        "operationId": "MenuItemController_findPopularItems",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "매장의 인기 메뉴를 반환합니다."
          }
        },
        "summary": "매장 인기 메뉴 조회",
        "tags": [
          "메뉴 아이템"
        ]
      }
    },
    "/api/menu-items/new/store/{storeId}": {
      "get": {
        "operationId": "MenuItemController_findNewItems",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "매장의 신메뉴를 반환합니다."
          }
        },
        "summary": "매장 신메뉴 조회",
        "tags": [
          "메뉴 아이템"
        ]
      }
    },
    "/api/menu-items/recommended/store/{storeId}": {
      "get": {
        "operationId": "MenuItemController_findRecommendedItems",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "매장의 추천 메뉴를 반환합니다."
          }
        },
        "summary": "매장 추천 메뉴 조회",
        "tags": [
          "메뉴 아이템"
        ]
      }
    },
...
    "/api/options/groups/store/{storeId}": {
      "post": {
        "operationId": "OptionController_createOptionGroup",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "description": "매장 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOptionGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "옵션 그룹 생성 성공"
          },
          "404": {
            "description": "매장을 찾을 수 없음"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "옵션 그룹 생성",
        "tags": [
          "옵션"
        ]
      }
    },
...
  },
  "info": {
    "title": "픽업 서버 API",
    "description": "픽업 서버 API 문서",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateMenuCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "커피",
            "description": "메뉴 카테고리 이름"
          },
          "description": {
            "type": "string",
            "example": "다양한 커피 음료",
            "description": "카테고리 설명"
          },
          "displayOrder": {
            "type": "number",
            "example": 1,
            "description": "카테고리 표시 순서"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "카테고리 활성 상태"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "CreateMenuItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "아메리카노",
            "description": "메뉴 이름"
          },
          "description": {
            "type": "string",
            "example": "진한 에스프레소에 물을 더한 음료",
            "description": "메뉴 설명"
          },
          "price": {
            "type": "number",
            "example": 4000,
            "description": "메뉴 가격"
          },
          "discountedPrice": {
            "type": "number",
            "example": 3500,
            "description": "할인된 가격"
          },
          "categoryId": {
            "type": "number",
            "example": 1,
            "description": "메뉴 카테고리 ID"
          },
          "image": {
            "type": "string",
            "description": "메뉴 이미지 파일",
            "format": "binary"
          },
          "preparationTime": {
            "type": "number",
            "example": 5,
            "description": "준비 시간 (분)"
          },
          "isPopular": {
            "type": "boolean",
            "example": true,
            "description": "인기 메뉴 여부"
          },
          "isNew": {
            "type": "boolean",
            "example": true,
            "description": "신메뉴 여부"
          },
          "isRecommended": {
            "type": "boolean",
            "example": true,
            "description": "추천 메뉴 여부"
          }
        },
        "required": [
          "name",
          "price"
        ]
      },
...
      "CreateOptionGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "옵션 그룹 이름"
          },
          "description": {
            "type": "string",
            "example": "음료 사이즈 선택",
            "description": "옵션 그룹 설명"
          },
          "isRequired": {
            "type": "boolean",
            "description": "필수 여부"
          },
          "minSelections": {
            "type": "number",
            "description": "최소 선택 개수"
          },
          "maxSelections": {
            "type": "number",
            "description": "최대 선택 개수"
          },
          "displayOrder": {
            "type": "number",
            "description": "표시 순서",
            "default": 0
          }
        },
        "required": [
          "name",
          "isRequired",
          "minSelections",
          "maxSelections"
        ]
      },
...
    }
  }
}